requirements.txt
asgiref
Django
django-filter
djangorestframework
gunicorn
packaging
psycopg2-binary
python-dotenv
pytz
sqlparse
typing_extensions
